Removed duplicate hooks and utils for query and mutation generation
