+++ packages/twenty-front/src/modules/object-record/cache/hooks/useUpsertFindManyRecordsQueryInCache.ts
@@ -5,21 +5,17 @@ import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadat
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
 import { MAX_QUERY_DEPTH_FOR_CACHE_INJECTION } from '@/object-record/cache/constants/MaxQueryDepthForCacheInjection';
 import { getRecordConnectionFromRecords } from '@/object-record/cache/utils/getRecordConnectionFromRecords';
-import { useGenerateFindManyRecordsQuery } from '@/object-record/hooks/useGenerateFindManyRecordsQuery';
 import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { ObjectRecordQueryVariables } from '@/object-record/types/ObjectRecordQueryVariables';
+import { generateFindManyRecordsQuery } from '@/object-record/utils/generateFindManyRecordsQuery';
 
 export const useUpsertFindManyRecordsQueryInCache = ({
   objectMetadataItem,
 }: {
-  objectMetadataItem: Pick<
-    ObjectMetadataItem,
-    'fields' | 'namePlural' | 'nameSingular'
-  >;
+  objectMetadataItem: ObjectMetadataItem;
 }) => {
   const apolloClient = useApolloClient();
 
-  const generateFindManyRecordsQuery = useGenerateFindManyRecordsQuery();
   const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
 
   const upsertFindManyRecordsQueryInCache = <
@@ -39,6 +35,7 @@ export const useUpsertFindManyRecordsQueryInCache = ({
   }) => {
     const findManyRecordsQueryForCacheOverwrite = generateFindManyRecordsQuery({
       objectMetadataItem,
+      objectMetadataItems,
       depth,
       queryFields,
       computeReferences,
