+++ packages/twenty-front/src/modules/object-record/constants/EmptyMutation.ts
