+++ packages/twenty-front/src/modules/object-record/constants/EmptyQuery.ts
