+++ packages/twenty-front/src/modules/object-record/hooks/useCreateManyRecordsMutation.ts
@@ -1,5 +1,13 @@
+import gql from 'graphql-tag';
+import { useRecoilValue } from 'recoil';
+
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useGenerateCreateManyRecordMutation } from '@/object-record/hooks/useGenerateCreateManyRecordMutation';
+import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
+import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
+import { EMPTY_MUTATION } from '@/object-record/constants/EmptyMutation';
+import { getCreateManyRecordsMutationResponseField } from '@/object-record/utils/getCreateManyRecordsMutationResponseField';
+import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
+import { capitalize } from '~/utils/string/capitalize';
 
 export const useCreateManyRecordsMutation = ({
   objectNameSingular,
@@ -14,11 +22,27 @@ export const useCreateManyRecordsMutation = ({
     objectNameSingular,
   });
 
-  const createManyRecordsMutation = useGenerateCreateManyRecordMutation({
-    objectMetadataItem,
-    depth,
-    queryFields,
-  });
+  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
+
+  if (isUndefinedOrNull(objectMetadataItem)) {
+    return { createManyRecordsMutation: EMPTY_MUTATION };
+  }
+
+  const mutationResponseField = getCreateManyRecordsMutationResponseField(
+    objectMetadataItem.namePlural,
+  );
+
+  const createManyRecordsMutation = gql`
+    mutation Create${capitalize(
+      objectMetadataItem.namePlural,
+    )}($data: [${capitalize(objectMetadataItem.nameSingular)}CreateInput!]!)  {
+      ${mutationResponseField}(data: $data) ${mapObjectMetadataToGraphQLQuery({
+        objectMetadataItems,
+        objectMetadataItem,
+        queryFields,
+        depth,
+      })}
+  }`;
 
   return {
     createManyRecordsMutation,
