+++ packages/twenty-front/src/modules/object-record/hooks/useCreateOneRecordMutation.ts
@@ -1,21 +1,49 @@
+import gql from 'graphql-tag';
+import { useRecoilValue } from 'recoil';
+
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useGenerateCreateOneRecordMutation } from '@/object-record/hooks/useGenerateCreateOneRecordMutation';
+import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
+import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
+import { EMPTY_MUTATION } from '@/object-record/constants/EmptyMutation';
+import { getCreateOneRecordMutationResponseField } from '@/object-record/utils/getCreateOneRecordMutationResponseField';
+import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
+import { capitalize } from '~/utils/string/capitalize';
 
 export const useCreateOneRecordMutation = ({
   objectNameSingular,
   queryFields,
+  depth,
 }: {
   objectNameSingular: string;
   queryFields?: Record<string, any>;
+  depth?: number;
 }) => {
   const { objectMetadataItem } = useObjectMetadataItem({
     objectNameSingular,
   });
 
-  const createOneRecordMutation = useGenerateCreateOneRecordMutation({
-    objectMetadataItem,
-    queryFields,
-  });
+  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
+
+  if (isUndefinedOrNull(objectMetadataItem)) {
+    return { createOneRecordMutation: EMPTY_MUTATION };
+  }
+
+  const capitalizedObjectName = capitalize(objectMetadataItem.nameSingular);
+
+  const mutationResponseField = getCreateOneRecordMutationResponseField(
+    objectMetadataItem.nameSingular,
+  );
+
+  const createOneRecordMutation = gql`
+    mutation CreateOne${capitalizedObjectName}($input: ${capitalizedObjectName}CreateInput!)  {
+      ${mutationResponseField}(data: $input) ${mapObjectMetadataToGraphQLQuery({
+        objectMetadataItems,
+        objectMetadataItem,
+        queryFields,
+        depth,
+      })}
+    }
+  `;
 
   return {
     createOneRecordMutation,
