+++ packages/twenty-front/src/modules/object-record/hooks/useDeleteOneRecordMutation.ts
@@ -1,5 +1,10 @@
+import gql from 'graphql-tag';
+
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { generateDeleteOneRecordMutation } from '@/object-record/utils/generateDeleteOneRecordMutation';
+import { EMPTY_MUTATION } from '@/object-record/constants/EmptyMutation';
+import { getDeleteOneRecordMutationResponseField } from '@/object-record/utils/getDeleteOneRecordMutationResponseField';
+import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
+import { capitalize } from '~/utils/string/capitalize';
 
 export const useDeleteOneRecordMutation = ({
   objectNameSingular,
@@ -10,9 +15,23 @@ export const useDeleteOneRecordMutation = ({
     objectNameSingular,
   });
 
-  const deleteOneRecordMutation = generateDeleteOneRecordMutation({
-    objectMetadataItem,
-  });
+  if (isUndefinedOrNull(objectMetadataItem)) {
+    return { deleteOneRecordMutation: EMPTY_MUTATION };
+  }
+
+  const capitalizedObjectName = capitalize(objectMetadataItem.nameSingular);
+
+  const mutationResponseField = getDeleteOneRecordMutationResponseField(
+    objectMetadataItem.nameSingular,
+  );
+
+  const deleteOneRecordMutation = gql`
+    mutation DeleteOne${capitalizedObjectName}($idToDelete: ID!)  {
+      ${mutationResponseField}(id: $idToDelete) {
+        id
+      }
+    }
+  `;
 
   return {
     deleteOneRecordMutation,
