+++ packages/twenty-front/src/modules/object-record/hooks/useFindManyRecordsQuery.ts
@@ -1,23 +1,33 @@
+import { useRecoilValue } from 'recoil';
+
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useGenerateFindManyRecordsQuery } from '@/object-record/hooks/useGenerateFindManyRecordsQuery';
+import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
 import { QueryFields } from '@/object-record/query-keys/types/QueryFields';
+import { generateFindManyRecordsQuery } from '@/object-record/utils/generateFindManyRecordsQuery';
 
 export const useFindManyRecordsQuery = ({
   objectNameSingular,
   queryFields,
+  depth,
+  computeReferences,
 }: {
   objectNameSingular: string;
   queryFields?: QueryFields;
+  depth?: number;
+  computeReferences?: boolean;
 }) => {
   const { objectMetadataItem } = useObjectMetadataItem({
     objectNameSingular,
   });
 
-  const generateFindManyRecordsQuery = useGenerateFindManyRecordsQuery();
+  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
 
   const findManyRecordsQuery = generateFindManyRecordsQuery({
     objectMetadataItem,
+    objectMetadataItems,
     queryFields,
+    depth,
+    computeReferences,
   });
 
   return {
