+++ packages/twenty-front/src/modules/object-record/hooks/useFindOneRecordQuery.ts
@@ -1,5 +1,10 @@
+import gql from 'graphql-tag';
+import { useRecoilValue } from 'recoil';
+
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useGenerateFindOneRecordQuery } from '@/object-record/hooks/useGenerateFindOneRecordQuery';
+import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
+import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
+import { capitalize } from '~/utils/string/capitalize';
 
 export const useFindOneRecordQuery = ({
   objectNameSingular,
@@ -12,12 +17,23 @@ export const useFindOneRecordQuery = ({
     objectNameSingular,
   });
 
-  const generateFindOneRecordQuery = useGenerateFindOneRecordQuery();
+  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
 
-  const findOneRecordQuery = generateFindOneRecordQuery({
-    objectMetadataItem,
-    depth,
-  });
+  const findOneRecordQuery = gql`
+      query FindOne${capitalize(
+        objectMetadataItem.nameSingular,
+      )}($objectRecordId: UUID!) {
+        ${objectMetadataItem.nameSingular}(filter: {
+          id: {
+            eq: $objectRecordId
+          }
+        })${mapObjectMetadataToGraphQLQuery({
+          objectMetadataItems,
+          objectMetadataItem,
+          depth,
+        })}
+      }
+  `;
 
   return {
     findOneRecordQuery,
