--- packages/twenty-front/src/modules/object-record/hooks/useGenerateCreateManyRecordMutation.ts
@@ -1,45 +1,0 @@
-import { gql } from '@apollo/client';
-import { useRecoilValue } from 'recoil';
-
-import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
-import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
-import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
-import { EMPTY_MUTATION } from '~/constants/EmptyMutation';
-import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
-import { capitalize } from '~/utils/string/capitalize';
-
-export const getCreateManyRecordsMutationResponseField = (
-  objectNamePlural: string,
-) => `create${capitalize(objectNamePlural)}`;
-
-export const useGenerateCreateManyRecordMutation = ({
-  objectMetadataItem,
-  queryFields,
-  depth = 1,
-}: {
-  objectMetadataItem: ObjectMetadataItem;
-  queryFields?: Record<string, any>;
-  depth?: number;
-}) => {
-  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
-
-  if (isUndefinedOrNull(objectMetadataItem)) {
-    return EMPTY_MUTATION;
-  }
-
-  const mutationResponseField = getCreateManyRecordsMutationResponseField(
-    objectMetadataItem.namePlural,
-  );
-
-  return gql`
-    mutation Create${capitalize(
-      objectMetadataItem.namePlural,
-    )}($data: [${capitalize(objectMetadataItem.nameSingular)}CreateInput!]!)  {
-      ${mutationResponseField}(data: $data) ${mapObjectMetadataToGraphQLQuery({
-        objectMetadataItems,
-        objectMetadataItem,
-        queryFields,
-        depth,
-      })}
-  }`;
-};
