--- packages/twenty-front/src/modules/object-record/hooks/useGenerateUpdateOneRecordMutation.ts
@@ -1,48 +1,0 @@
-import { gql } from '@apollo/client';
-import { useRecoilValue } from 'recoil';
-
-import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
-import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
-import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
-import { EMPTY_MUTATION } from '~/constants/EmptyMutation';
-import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
-import { capitalize } from '~/utils/string/capitalize';
-
-export const getUpdateOneRecordMutationResponseField = (
-  objectNameSingular: string,
-) => `update${capitalize(objectNameSingular)}`;
-
-export const useGenerateUpdateOneRecordMutation = ({
-  objectMetadataItem,
-  depth = 1,
-  computeReferences = false,
-}: {
-  objectMetadataItem: ObjectMetadataItem;
-  depth?: number;
-  computeReferences?: boolean;
-}) => {
-  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
-
-  if (isUndefinedOrNull(objectMetadataItem)) {
-    return EMPTY_MUTATION;
-  }
-
-  const capitalizedObjectName = capitalize(objectMetadataItem.nameSingular);
-
-  const mutationResponseField = getUpdateOneRecordMutationResponseField(
-    objectMetadataItem.nameSingular,
-  );
-
-  return gql`
-    mutation UpdateOne${capitalizedObjectName}($idToUpdate: ID!, $input: ${capitalizedObjectName}UpdateInput!)  {
-       ${mutationResponseField}(id: $idToUpdate, data: $input) ${mapObjectMetadataToGraphQLQuery(
-         {
-           objectMetadataItems,
-           objectMetadataItem,
-           depth,
-           computeReferences,
-         },
-       )}
-    }
-  `;
-};
