+++ packages/twenty-front/src/modules/object-record/hooks/useUpdateOneRecordMutation.ts
@@ -1,21 +1,51 @@
+import gql from 'graphql-tag';
+import { useRecoilValue } from 'recoil';
+
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useGenerateUpdateOneRecordMutation } from '@/object-record/hooks/useGenerateUpdateOneRecordMutation';
+import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
+import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
+import { EMPTY_MUTATION } from '@/object-record/constants/EmptyMutation';
+import { getUpdateOneRecordMutationResponseField } from '@/object-record/utils/getUpdateOneRecordMutationResponseField';
+import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
+import { capitalize } from '~/utils/string/capitalize';
 
 export const useUpdateOneRecordMutation = ({
   objectNameSingular,
   computeReferences = false,
+  depth,
 }: {
   objectNameSingular: string;
   computeReferences?: boolean;
+  depth?: number;
 }) => {
   const { objectMetadataItem } = useObjectMetadataItem({
     objectNameSingular,
   });
 
-  const updateOneRecordMutation = useGenerateUpdateOneRecordMutation({
-    objectMetadataItem,
-    computeReferences,
-  });
+  const objectMetadataItems = useRecoilValue(objectMetadataItemsState);
+
+  if (isUndefinedOrNull(objectMetadataItem)) {
+    return { updateOneRecordMutation: EMPTY_MUTATION };
+  }
+
+  const capitalizedObjectName = capitalize(objectMetadataItem.nameSingular);
+
+  const mutationResponseField = getUpdateOneRecordMutationResponseField(
+    objectMetadataItem.nameSingular,
+  );
+
+  const updateOneRecordMutation = gql`
+    mutation UpdateOne${capitalizedObjectName}($idToUpdate: ID!, $input: ${capitalizedObjectName}UpdateInput!)  {
+       ${mutationResponseField}(id: $idToUpdate, data: $input) ${mapObjectMetadataToGraphQLQuery(
+         {
+           objectMetadataItems,
+           objectMetadataItem,
+           depth,
+           computeReferences,
+         },
+       )}
+    }
+  `;
 
   return {
     updateOneRecordMutation,
