+++ packages/twenty-front/src/modules/object-record/utils/getCreateManyRecordsMutationResponseField.ts
@@ -1,0 +1,5 @@
+import { capitalize } from '~/utils/string/capitalize';
+
+export const getCreateManyRecordsMutationResponseField = (
+  objectNamePlural: string,
+) => `create${capitalize(objectNamePlural)}`;
