+++ packages/twenty-front/src/modules/object-record/utils/getDeleteManyRecordsMutationResponseField.ts
@@ -1,0 +1,5 @@
+import { capitalize } from '~/utils/string/capitalize';
+
+export const getDeleteManyRecordsMutationResponseField = (
+  objectNamePlural: string,
+) => `delete${capitalize(objectNamePlural)}`;
