+++ packages/twenty-front/src/modules/object-record/utils/getDeleteOneRecordMutationResponseField.ts
@@ -1,0 +1,5 @@
+import { capitalize } from '~/utils/string/capitalize';
+
+export const getDeleteOneRecordMutationResponseField = (
+  objectNameSingular: string,
+) => `delete${capitalize(objectNameSingular)}`;
