+++ packages/twenty-front/src/modules/object-record/utils/getFindDuplicateRecordsQueryResponseField.ts
@@ -1,0 +1,3 @@
+export const getFindDuplicateRecordsQueryResponseField = (
+  objectNameSingular: string,
+) => `${objectNameSingular}Duplicates`;
