+++ packages/twenty-front/src/modules/object-record/utils/getUpdateOneRecordMutationResponseField.ts
@@ -1,0 +1,5 @@
+import { capitalize } from '~/utils/string/capitalize';
+
+export const getUpdateOneRecordMutationResponseField = (
+  objectNameSingular: string,
+) => `update${capitalize(objectNameSingular)}`;
